[Cursor] 배당률 소수점 3자리 정밀도 통일 및 예상수익 계산 오차 수정
- 모든 배당률 표시를 toFixed(2)에서 toFixed(3)으로 변경
- ExchangeContext에서 배당률 3자리 floor 처리로 정확한 계산 보장
- BetSelectionPanel과 ExchangeSidebar의 예상수익 계산 부동소수점 오차 방지
- 백엔드 모델의 DECIMAL 정밀도 3자리로 통일
- oddsUtils.js의 adjustOddsPayout 함수 3자리 floor 처리 적용

+++ server/models/exchangeOrderModel.js
@@ -17,7 +17,7 @@ const ExchangeOrder = sequelize.define('ExchangeOrder', {
   market: { type: DataTypes.STRING, allowNull: false },
   line: { type: DataTypes.FLOAT, allowNull: false },
   side: { type: DataTypes.ENUM('back', 'lay'), allowNull: false },
-  price: { type: DataTypes.FLOAT, allowNull: false },
+  price: { type: DataTypes.DECIMAL(10, 3), allowNull: false },
   amount: { type: DataTypes.INTEGER, allowNull: false },
   selection: { type: DataTypes.STRING, allowNull: true }, // 선택한 팀/선수명 (기존 호환성)
   status: { 
@@ -111,7 +111,7 @@ const ExchangeOrder = sequelize.define('ExchangeOrder', {
     comment: '멀티배팅 여부'
   },
   totalOdds: {
-    type: DataTypes.DECIMAL(10, 4),
+    type: DataTypes.DECIMAL(10, 3),
     allowNull: true,
     comment: '총 배당률 (멀티배팅용)'
   },
